modify: 각 코드를 별도의 파일로 분리
