+++ first-example/index.js
@@ -1,10 +1,11 @@
 // 공연료 청구서 출력
 const invoice = require('./data/invoices.json')
 const play = require('./data/plays.json')
+const createStatementData = require('./createStatementData.js')
 
 console.log(statement(invoice, play))
 
-function renderPlainText(data, plays) {
+function renderPlainText(data) {
   let result = `청구내역 (고객명: ${data.customer})\n`
 
   for (let perf of data.performances) {
@@ -27,67 +28,6 @@ function renderPlainText(data, plays) {
 
 }
 
-function createStatementData(invoice, plays) {
-  const statementData = {}
-  statementData.customer = invoice.customer
-  statementData.performances = invoice.performances.map(enrichPerfomence)
-  statementData.totalAmount = totalAmount(statementData)
-  statementData.totalVolumeCredits = totalVolumeCredits(statementData)
-  return statementData
-
-
-  function totalVolumeCredits(data) {
-    return data.performances.reduce((total, p) => total + p.volumeCredits, 0)
-  }
-
-  function totalAmount(data) {
-    return data.performances.reduce((total, p) => total + p.amount, 0)
-  }
-
-  function enrichPerfomence(aPerformance) {
-    const result = Object.assign({}, aPerformance) // 얕은 복사
-    result.play = playFor(result)
-    result.amount = amountFor(result)
-    result.volumeCredits = volumeCreditsFor(result)
-    return result
-  }
-
-  function playFor(aPerformance) {
-    return plays[aPerformance.playID]
-  }
-
-  function amountFor(aPerformance) {
-    let result = 0
-    switch (aPerformance.play.type) {
-      case 'tragedy':
-        result = 40000
-        if (aPerformance.audience > 30) {
-          result += 1000 * (aPerformance.audience - 30)
-        }
-        break
-      case 'comedy':
-        result = 30000
-        if (aPerformance.audience > 20) {
-          result += 10000 + 500 * (aPerformance.audience - 20)
-        }
-        result += 300 * aPerformance.audience
-        break
-      default:
-        throw new Error(`알 수 없는 장르: ${aPerformance.play.type}`)
-    }
-
-    return result
-  }
-
-  function volumeCreditsFor(aPerformance) {
-    let result = 0
-    result += Math.max(aPerformance.audience - 30, 0)
-    if ('comedy' === aPerformance.play.type)
-      result += Math.floor(aPerformance.audience / 5)
-
-    return result
-  }
-}
 
 function statement(invoice, plays) {
   return renderPlainText(createStatementData(invoice, plays))
